refactor(gui): improve widget update processing
BREAKING CHANGE: Some widget operation functions have been renamed or removed.
