+++ src/gui/widget_style.c
@@ -248,12 +248,12 @@ void Widget_AddTaskByStyle(LCUI_Widget w, int key)
 		{ key_opacity, key_opacity, LCUI_WTASK_OPACITY, TRUE },
 		{ key_z_index, key_z_index, LCUI_WTASK_ZINDEX, TRUE },
 		{ key_width, key_height, LCUI_WTASK_RESIZE, TRUE },
-		{ key_min_width, key_max_height, LCUI_WTASK_RESIZE_WITH_SURFACE,
+		{ key_min_width, key_max_height, LCUI_WTASK_RESIZE,
 		  TRUE },
 		{ key_padding_start, key_padding_end,
-		  LCUI_WTASK_RESIZE_WITH_SURFACE, TRUE },
+		  LCUI_WTASK_RESIZE, TRUE },
 		{ key_box_sizing, key_box_sizing,
-		  LCUI_WTASK_RESIZE_WITH_SURFACE, TRUE },
+		  LCUI_WTASK_RESIZE, TRUE },
 		{ key_margin_start, key_margin_end, LCUI_WTASK_MARGIN, TRUE },
 		{ key_position_start, key_position_end, LCUI_WTASK_POSITION,
 		  TRUE },
