+++ src/gui/widget_task.c
@@ -1,7 +1,7 @@
 ﻿/*
  * widget_task.c -- LCUI widget task module.
  *
- * Copyright (c) 2018, Liu chao <lc-soft@live.cn> All rights reserved.
+ * Copyright (c) 2018-2019, Liu chao <lc-soft@live.cn> All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -30,17 +30,54 @@
 
 #include <time.h>
 #include <stdio.h>
+#include <string.h>
 #include <stdlib.h>
 #include <LCUI_Build.h>
 #include <LCUI/LCUI.h>
 #include <LCUI/gui/widget.h>
+#include <LCUI/gui/metrics.h>
+
+#define MEMCMP(A, B) memcmp(A, B, sizeof(*(A)))
+
+typedef struct LCUI_WidgetTaskContextRec_ *LCUI_WidgetTaskContext;
+
+/** for check widget difference */
+typedef struct LCUI_WidgetTaskDiffRec_ {
+	int z_index;
+	int display;
+	float opacity;
+	LCUI_BOOL visible;
+	LCUI_StyleValue position;
+	LCUI_BorderStyle border;
+	LCUI_BoxShadowStyle shadow;
+	LCUI_BackgroundStyle background;
+	LCUI_WidgetBoxModelRec box;
+
+	int invalid_box;
+	LCUI_BOOL can_render;
+	LCUI_BOOL sync_props_to_surface;
+	LCUI_BOOL should_add_invalid_area;
+} LCUI_WidgetTaskDiffRec, *LCUI_WidgetTaskDiff;
+
+typedef struct LCUI_WidgetTaskContextRec_ {
+	unsigned style_hash;
+	Dict *style_cache;
+
+	LCUI_WidgetTaskDiffRec diff;
+	LCUI_WidgetTaskContext parent;
+	LCUI_WidgetTasksProfile profile;
+} LCUI_WidgetTaskContextRec;
 
 static struct WidgetTaskModule {
-	DictType style_cache_dict;
 	unsigned max_updates_per_frame;
+	DictType style_cache_dict;
+	LCUI_MetricsRec metrics;
+	LCUI_BOOL refresh_all;
 	LCUI_WidgetFunction handlers[LCUI_WTASK_TOTAL_NUM];
-	unsigned update_count;
 } self;
+
+static size_t Widget_UpdateWithContext(LCUI_Widget w,
+				       LCUI_WidgetTaskContext ctx);
 
 static unsigned int IntKeyDict_HashFunction(const void *key)
 {
@@ -83,7 +120,99 @@ static void InitStylesheetCacheDict(void)
 	dt->keyDestructor = IntKeyDict_KeyDestructor;
 }
 
-static void HandleRefreshStyle(LCUI_Widget w)
+static int ComputeStyleOption(LCUI_Widget w, int key, int default_value)
+{
+	if (!w->style->sheet[key].is_valid) {
+		return default_value;
+	}
+	if (w->style->sheet[key].type != LCUI_STYPE_STYLE) {
+		return default_value;
+	}
+	return w->style->sheet[key].style;
+}
+
+static void Widget_ComputePaddingStyle(LCUI_Widget w)
+{
+	int i;
+	LCUI_Style s;
+	LCUI_BoundBox *pbox = &w->computed_style.padding;
+	struct {
+		LCUI_Style sval;
+		float *ival;
+		int key;
+	} pd_map[4] = { { &pbox->top, &w->padding.top, key_padding_top },
+			{ &pbox->right, &w->padding.right, key_padding_right },
+			{ &pbox->bottom, &w->padding.bottom,
+			  key_padding_bottom },
+			{ &pbox->left, &w->padding.left, key_padding_left } };
+
+	/* 内边距的单位暂时都用 px  */
+	for (i = 0; i < 4; ++i) {
+		s = &w->style->sheet[pd_map[i].key];
+		if (!s->is_valid) {
+			pd_map[i].sval->type = LCUI_STYPE_PX;
+			pd_map[i].sval->px = 0.0;
+			*pd_map[i].ival = 0.0;
+			continue;
+		}
+		*pd_map[i].sval = *s;
+		*pd_map[i].ival = LCUIMetrics_Compute(s->value, s->type);
+	}
+}
+
+static void Widget_ComputeMarginStyle(LCUI_Widget w)
+{
+	int i;
+	LCUI_Style s;
+	LCUI_BoundBox *mbox = &w->computed_style.margin;
+	struct {
+		LCUI_Style sval;
+		float *fval;
+		int key;
+	} pd_map[4] = { { &mbox->top, &w->margin.top, key_margin_top },
+			{ &mbox->right, &w->margin.right, key_margin_right },
+			{ &mbox->bottom, &w->margin.bottom, key_margin_bottom },
+			{ &mbox->left, &w->margin.left, key_margin_left } };
+
+	for (i = 0; i < 4; ++i) {
+		s = &w->style->sheet[pd_map[i].key];
+		if (!s->is_valid) {
+			pd_map[i].sval->type = LCUI_STYPE_PX;
+			pd_map[i].sval->px = 0.0;
+			*pd_map[i].fval = 0.0;
+			continue;
+		}
+		*pd_map[i].sval = *s;
+		*pd_map[i].fval = LCUIMetrics_Compute(s->value, s->type);
+	}
+	/* 如果有父级部件，则处理 margin-left 和 margin-right 的值 */
+	if (w->parent) {
+		float width = w->parent->box.content.width;
+		if (Widget_HasAutoStyle(w, key_margin_left)) {
+			if (Widget_HasAutoStyle(w, key_margin_right)) {
+				w->margin.left = (width - w->width) / 2;
+				if (w->margin.left < 0) {
+					w->margin.left = 0;
+				}
+				w->margin.right = w->margin.left;
+			} else {
+				w->margin.left = width - w->width;
+				w->margin.left -= w->margin.right;
+				if (w->margin.left < 0) {
+					w->margin.left = 0;
+				}
+			}
+		} else if (Widget_HasAutoStyle(w, key_margin_right)) {
+			w->margin.right = width - w->width;
+			w->margin.right -= w->margin.left;
+			if (w->margin.right < 0) {
+				w->margin.right = 0;
+			}
+		}
+	}
+}
+
+static void Widget_OnRefreshStyle(LCUI_Widget w)
 {
 	int i;
 
@@ -94,32 +223,327 @@ static void HandleRefreshStyle(LCUI_Widget w)
 	w->task.states[LCUI_WTASK_UPDATE_STYLE] = FALSE;
 }
 
-static void HandleUpdateStyle(LCUI_Widget w)
+static void Widget_OnUpdateStyle(LCUI_Widget w)
 {
 	Widget_ExecUpdateStyle(w, FALSE);
 }
 
-static void HandleSetTitle(LCUI_Widget w)
+static void Widget_OnSetTitle(LCUI_Widget w)
 {
 	Widget_PostSurfaceEvent(w, LCUI_WEVENT_TITLE, TRUE);
 }
 
-/** 处理主体刷新（标记主体区域为脏矩形，但不包括阴影区域） */
-static void HandleBody(LCUI_Widget w)
+static void Widget_ComputeFlexLayoutStyle(LCUI_Widget w)
 {
-	Widget_InvalidateArea(w, NULL, SV_BORDER_BOX);
+	LCUI_FlexLayoutStyle *data = &w->computed_style.flex;
+	LCUI_Style s = StyleSheet_GetStyle(w->style, key_justify_content);
+
+	if (s->type != LCUI_STYPE_STYLE || !s->is_valid) {
+		data->justify_content = SV_FLEX_START;
+		return;
+	}
+	data->justify_content = s->val_style;
 }
 
-/** 处理刷新（标记整个部件区域为脏矩形） */
-static void HandleRefresh(LCUI_Widget w)
+static void Widget_OnUpdateBorder(LCUI_Widget w)
 {
-	DEBUG_MSG("refresh\n");
-	Widget_InvalidateArea(w, NULL, SV_GRAPH_BOX);
+	Widget_ComputeBorderStyle(w);
+}
+
+static void Widget_OnUpdateVisible(LCUI_Widget w)
+{
+	LCUI_Style s = &w->style->sheet[key_visibility];
+
+	if (w->computed_style.display == SV_NONE) {
+		w->computed_style.visible = FALSE;
+	} else if (s->is_valid && s->type == LCUI_STYPE_STRING &&
+		   strcmp(s->val_string, "hidden") == 0) {
+		w->computed_style.visible = FALSE;
+	} else {
+		w->computed_style.visible = TRUE;
+	}
+}
+
+static void Widget_OnUpdateDisplay(LCUI_Widget w)
+{
+	LCUI_Style s = &w->style->sheet[key_display];
+	LCUI_WidgetStyle *style = &w->computed_style;
+
+	if (s->is_valid && s->type == LCUI_STYPE_STYLE) {
+		style->display = s->style;
+		if (style->display == SV_NONE) {
+			w->computed_style.visible = FALSE;
+		}
+	} else {
+		style->display = SV_BLOCK;
+	}
+	Widget_OnUpdateVisible(w);
+}
+
+static void Widget_OnUpdateOpacity(LCUI_Widget w)
+{
+	float opacity = 1.0;
+	LCUI_Style s = &w->style->sheet[key_opacity];
+
+	if (s->is_valid) {
+		switch (s->type) {
+		case LCUI_STYPE_INT:
+			opacity = 1.0f * s->val_int;
+			break;
+		case LCUI_STYPE_SCALE:
+			opacity = s->val_scale;
+			break;
+		default:
+			opacity = 1.0f;
+			break;
+		}
+		if (opacity > 1.0) {
+			opacity = 1.0;
+		} else if (opacity < 0.0) {
+			opacity = 0.0;
+		}
+	}
+	w->computed_style.opacity = opacity;
+}
+
+static void Widget_OnUpdateZIndex(LCUI_Widget w)
+{
+	LCUI_Style s = &w->style->sheet[key_z_index];
+
+	if (s->is_valid && s->type == LCUI_STYPE_INT) {
+		w->computed_style.z_index = s->val_int;
+	} else {
+		w->computed_style.z_index = 0;
+	}
+}
+
+static void Widget_ClearComputedSize(LCUI_Widget w)
+{
+	if (Widget_HasAutoStyle(w, key_width)) {
+		w->width = 0;
+		w->box.canvas.width = 0;
+		w->box.content.width = 0;
+	}
+	if (Widget_HasAutoStyle(w, key_height)) {
+		w->height = 0;
+		w->box.canvas.height = 0;
+		w->box.content.height = 0;
+	}
+}
+
+static void Widget_UpdateChildrenSize(LCUI_Widget w)
+{
+	LinkedListNode *node;
+
+	for (LinkedList_Each(node, &w->children)) {
+		LCUI_Widget child = node->data;
+		LCUI_StyleSheet s = child->style;
+
+		if (Widget_HasFillAvailableWidth(child)) {
+			Widget_AddTask(child, LCUI_WTASK_RESIZE);
+		} else if (Widget_HasScaleSize(child)) {
+			Widget_AddTask(child, LCUI_WTASK_RESIZE);
+		}
+		if (Widget_HasAbsolutePosition(child)) {
+			if (s->sheet[key_right].is_valid ||
+			    s->sheet[key_bottom].is_valid ||
+			    CheckStyleType(s, key_left, scale) ||
+			    CheckStyleType(s, key_top, scale)) {
+				Widget_AddTask(child, LCUI_WTASK_POSITION);
+			}
+		}
+		if (Widget_HasAutoStyle(child, key_margin_left) ||
+		    Widget_HasAutoStyle(child, key_margin_right)) {
+			Widget_AddTask(child, LCUI_WTASK_MARGIN);
+		}
+		if (child->computed_style.vertical_align != SV_TOP) {
+			Widget_AddTask(child, LCUI_WTASK_POSITION);
+		}
+	}
+}
+
+static void Widget_OnUpdatePosition(LCUI_Widget w)
+{
+	int position = ComputeStyleOption(w, key_position, SV_STATIC);
+	int valign = ComputeStyleOption(w, key_vertical_align, SV_TOP);
+
+	w->computed_style.vertical_align = valign;
+	w->computed_style.left = Widget_ComputeXMetric(w, key_left);
+	w->computed_style.right = Widget_ComputeXMetric(w, key_right);
+	w->computed_style.top = Widget_ComputeYMetric(w, key_top);
+	w->computed_style.bottom = Widget_ComputeYMetric(w, key_bottom);
+	if (w->parent && w->computed_style.position != position) {
+		w->computed_style.position = position;
+		Widget_AddTask(w->parent, LCUI_WTASK_LAYOUT);
+		Widget_ClearComputedSize(w);
+		Widget_UpdateChildrenSize(w);
+		/* 当部件尺寸是按百分比动态计算的时候需要重新计算尺寸 */
+		if (Widget_CheckStyleType(w, key_width, scale) ||
+		    Widget_CheckStyleType(w, key_height, scale)) {
+			Widget_AddTask(w, LCUI_WTASK_RESIZE);
+		}
+	}
+	w->computed_style.position = position;
+	Widget_OnUpdateZIndex(w);
+	w->x = w->origin_x;
+	w->y = w->origin_y;
+	switch (position) {
+	case SV_ABSOLUTE:
+		w->x = w->y = 0;
+		if (!Widget_HasAutoStyle(w, key_left)) {
+			w->x = w->computed_style.left;
+		} else if (!Widget_HasAutoStyle(w, key_right)) {
+			if (w->parent) {
+				w->x = w->parent->box.border.width;
+				w->x -= w->width;
+			}
+			w->x -= w->computed_style.right;
+		}
+		if (!Widget_HasAutoStyle(w, key_top)) {
+			w->y = w->computed_style.top;
+		} else if (!Widget_HasAutoStyle(w, key_bottom)) {
+			if (w->parent) {
+				w->y = w->parent->box.border.height;
+				w->y -= w->height;
+			}
+			w->y -= w->computed_style.bottom;
+		}
+		break;
+	case SV_RELATIVE:
+		if (!Widget_HasAutoStyle(w, key_left)) {
+			w->x += w->computed_style.left;
+		} else if (!Widget_HasAutoStyle(w, key_right)) {
+			w->x -= w->computed_style.right;
+		}
+		if (!Widget_HasAutoStyle(w, key_top)) {
+			w->y += w->computed_style.top;
+		} else if (!Widget_HasAutoStyle(w, key_bottom)) {
+			w->y -= w->computed_style.bottom;
+		}
+	default:
+		if (w->parent) {
+			w->x += w->parent->padding.left;
+			w->y += w->parent->padding.top;
+		}
+		break;
+	}
+	switch (valign) {
+	case SV_MIDDLE:
+		if (!w->parent) {
+			break;
+		}
+		w->y += (w->parent->box.content.height - w->height) / 2;
+		break;
+	case SV_BOTTOM:
+		if (!w->parent) {
+			break;
+		}
+		w->y += w->parent->box.content.height - w->height;
+	case SV_TOP:
+	default:
+		break;
+	}
+	w->box.outer.x = w->x;
+	w->box.outer.y = w->y;
+	w->x += w->margin.left;
+	w->y += w->margin.top;
+	/* 以x、y为基础 */
+	w->box.padding.x = w->x;
+	w->box.padding.y = w->y;
+	w->box.border.x = w->x;
+	w->box.border.y = w->y;
+	w->box.canvas.x = w->x;
+	w->box.canvas.y = w->y;
+	/* 计算各个框的坐标 */
+	w->box.padding.x += w->computed_style.border.left.width;
+	w->box.padding.y += w->computed_style.border.top.width;
+	w->box.content.x = w->box.padding.x + w->padding.left;
+	w->box.content.y = w->box.padding.y + w->padding.top;
+	w->box.canvas.x -= Widget_GetBoxShadowOffsetX(w);
+	w->box.canvas.y -= Widget_GetBoxShadowOffsetY(w);
+}
+
+static void Widget_OnUpdateMargin(LCUI_Widget w)
+{
+	Widget_ComputeMarginStyle(w);
+}
+
+static void Widget_OnUpdatePadding(LCUI_Widget w)
+{
+	Widget_ComputePaddingStyle(w);
+}
+
+static void Widget_OnUpdateSize(LCUI_Widget w)
+{
+	int box_sizing;
+	LCUI_RectF rect = w->box.canvas;
+
+	box_sizing = ComputeStyleOption(w, key_box_sizing, SV_CONTENT_BOX);
+	w->computed_style.box_sizing = box_sizing;
+	Widget_ComputePaddingStyle(w);
+	Widget_ComputeSizeStyle(w);
+	/* 如果左右外间距是 auto 类型的，则需要计算外间距 */
+	if (w->style->sheet[key_margin_left].is_valid &&
+	    w->style->sheet[key_margin_left].type == LCUI_STYPE_AUTO) {
+		Widget_ComputeMarginStyle(w);
+	} else if (w->style->sheet[key_margin_right].is_valid &&
+		   w->style->sheet[key_margin_right].type == LCUI_STYPE_AUTO) {
+		Widget_ComputeMarginStyle(w);
+	}
+	/* 若在变化前后的宽高中至少有一个为 0，则不继续处理 */
+	if ((w->box.canvas.width <= 0 || w->box.canvas.height <= 0) &&
+	    (rect.width <= 0 || rect.height <= 0)) {
+		return;
+	}
+	/* 如果垂直对齐方式不为顶部对齐 */
+	if (w->computed_style.vertical_align != SV_TOP) {
+		Widget_AddTask(w, LCUI_WTASK_POSITION);
+	} else if (w->computed_style.position == SV_ABSOLUTE) {
+		/* 如果是绝对定位，且指定了右间距或底间距 */
+		if (!Widget_HasAutoStyle(w, key_right) ||
+		    !Widget_HasAutoStyle(w, key_bottom)) {
+			Widget_AddTask(w, LCUI_WTASK_POSITION);
+		}
+	}
+}
+
+static void Widget_OnUpdateBoxShadow(LCUI_Widget w)
+{
+	Widget_ComputeBoxShadowStyle(w);
+}
+
+static void Widget_OnUpdateBackground(LCUI_Widget w)
+{
+	Widget_ComputeBackgroundStyle(w);
+}
+
+static void Widget_OnUpdateLayout(LCUI_Widget w)
+{
+	if (w->computed_style.display == SV_FLEX) {
+		Widget_ComputeFlexLayoutStyle(w);
+	}
+	Widget_DoLayout(w);
+}
+
+static void Widget_OnUpdateProps(LCUI_Widget w)
+{
+	LCUI_Style s;
+	LCUI_WidgetStyle *style = &w->computed_style;
+
+	s = &w->style->sheet[key_focusable];
+	style->pointer_events =
+	    ComputeStyleOption(w, key_pointer_events, SV_INHERIT);
+	if (s->is_valid && s->type == LCUI_STYPE_BOOL && s->val_bool == 0) {
+		style->focusable = FALSE;
+	} else {
+		style->focusable = TRUE;
+	}
 }
 
 void Widget_UpdateTaskStatus(LCUI_Widget widget)
 {
 	int i;
+
 	for (i = 0; i < LCUI_WTASK_TOTAL_NUM; ++i) {
 		if (widget->task.states[i]) {
 			break;
@@ -165,32 +589,186 @@ void Widget_AddTask(LCUI_Widget widget, int task)
 
 void LCUIWidget_InitTasks(void)
 {
-#define SetHandler(NAME, HANDLER) self.handlers[LCUI_WTASK_##NAME] = HANDLER
-	SetHandler(VISIBLE, Widget_UpdateVisibility);
-	SetHandler(POSITION, Widget_UpdatePosition);
-	SetHandler(RESIZE, Widget_UpdateSize);
-	SetHandler(RESIZE_WITH_SURFACE, Widget_UpdateSizeWithSurface);
-	SetHandler(SHADOW, Widget_UpdateBoxShadow);
-	SetHandler(BORDER, Widget_UpdateBorder);
-	SetHandler(OPACITY, Widget_UpdateOpacity);
-	SetHandler(MARGIN, Widget_UpdateMargin);
-	SetHandler(BODY, HandleBody);
-	SetHandler(TITLE, HandleSetTitle);
-	SetHandler(REFRESH, HandleRefresh);
-	SetHandler(UPDATE_STYLE, HandleUpdateStyle);
-	SetHandler(REFRESH_STYLE, HandleRefreshStyle);
-	SetHandler(BACKGROUND, Widget_UpdateBackground);
-	SetHandler(LAYOUT, Widget_ExecUpdateLayout);
-	SetHandler(ZINDEX, Widget_ExecUpdateZIndex);
-	SetHandler(DISPLAY, Widget_UpdateDisplay);
-	SetHandler(PROPS, Widget_UpdateProps);
+#define SetHandler(NAME, HANDLER) \
+	self.handlers[LCUI_WTASK_##NAME] = Widget_On##HANDLER
+	SetHandler(VISIBLE, UpdateVisible);
+	SetHandler(POSITION, UpdatePosition);
+	SetHandler(RESIZE, UpdateSize);
+	SetHandler(SHADOW, UpdateBoxShadow);
+	SetHandler(BORDER, UpdateBorder);
+	SetHandler(OPACITY, UpdateOpacity);
+	SetHandler(MARGIN, UpdateMargin);
+	SetHandler(PADDING, UpdatePadding);
+	SetHandler(TITLE, SetTitle);
+	SetHandler(UPDATE_STYLE, UpdateStyle);
+	SetHandler(REFRESH_STYLE, RefreshStyle);
+	SetHandler(BACKGROUND, UpdateBackground);
+	SetHandler(LAYOUT, UpdateLayout);
+	SetHandler(ZINDEX, UpdateZIndex);
+	SetHandler(DISPLAY, UpdateDisplay);
+	SetHandler(PROPS, UpdateProps);
 	InitStylesheetCacheDict();
+	self.refresh_all = TRUE;
 	self.max_updates_per_frame = 4;
 }
 
 void LCUIWidget_FreeTasks(void)
 {
 	LCUIWidget_ClearTrash();
+}
+
+static void Widget_InitDiff(LCUI_Widget w, LCUI_WidgetTaskContext ctx)
+{
+	ctx->diff.can_render = TRUE;
+	ctx->diff.invalid_box = self.refresh_all ? SV_GRAPH_BOX : 0;
+	ctx->diff.should_add_invalid_area = FALSE;
+	if (ctx->parent) {
+		if (!ctx->parent->diff.can_render) {
+			ctx->diff.can_render = FALSE;
+			return;
+		}
+		if (ctx->parent->diff.invalid_box >= SV_PADDING_BOX) {
+			ctx->diff.invalid_box = SV_GRAPH_BOX;
+			return;
+		}
+	}
+	if (w->state < LCUI_WSTATE_LAYOUTED) {
+		ctx->diff.invalid_box = SV_GRAPH_BOX;
+	}
+	ctx->diff.should_add_invalid_area = TRUE;
+}
+
+static void Widget_BeginDiff(LCUI_Widget w, LCUI_WidgetTaskContext ctx)
+{
+	const LCUI_WidgetStyle *style = &w->computed_style;
+
+	if (self.refresh_all) {
+		memset(&ctx->diff, 0, sizeof(ctx->diff));
+		Widget_InitDiff(w, ctx);
+	} else {
+		ctx->diff.box = w->box;
+		ctx->diff.display = style->display;
+		ctx->diff.z_index = style->z_index;
+		ctx->diff.visible = style->visible;
+		ctx->diff.opacity = style->opacity;
+		ctx->diff.position = style->position;
+		ctx->diff.shadow = style->shadow;
+		ctx->diff.border = style->border;
+		ctx->diff.background = style->background;
+	}
+}
+
+static int Widget_EndDiff(LCUI_Widget w, LCUI_WidgetTaskContext ctx)
+{
+	LCUI_RectF rect;
+	LCUI_WidgetEventRec e;
+	LCUI_BOOL widget_box_moved = FALSE;
+	const LCUI_WidgetTaskDiff diff = &ctx->diff;
+	const LCUI_WidgetStyle *style = &w->computed_style;
+
+	if (!diff->can_render) {
+		return 0;
+	}
+	diff->can_render = style->visible;
+	/* hidden widget do not need to be rendered */
+	if (style->visible == diff->visible && !style->visible) {
+		return 0;
+	}
+	/* Check for layout related property changes */
+
+	if (diff->visible != style->visible) {
+		diff->invalid_box = SV_GRAPH_BOX;
+		if (style->visible) {
+			Widget_PostSurfaceEvent(w, LCUI_WEVENT_SHOW, TRUE);
+		} else {
+			Widget_PostSurfaceEvent(w, LCUI_WEVENT_HIDE, TRUE);
+		}
+	}
+	if (style->position == SV_ABSOLUTE) {
+		if (MEMCMP(&diff->box.canvas, &w->box.canvas)) {
+			diff->invalid_box = SV_GRAPH_BOX;
+			widget_box_moved = TRUE;
+		}
+	} else if (MEMCMP(&diff->box.outer, &w->box.outer)) {
+		diff->invalid_box = SV_GRAPH_BOX;
+		if (w->parent) {
+			Widget_AddTask(w->parent, LCUI_WTASK_LAYOUT);
+		}
+		Widget_PostSurfaceEvent(w, LCUI_WEVENT_MOVE,
+					!w->task.skip_surface_props_sync);
+		w->task.skip_surface_props_sync = TRUE;
+		widget_box_moved = TRUE;
+	} else if (MEMCMP(&diff->box.canvas, &w->box.canvas)) {
+		diff->invalid_box = SV_GRAPH_BOX;
+		widget_box_moved = TRUE;
+	}
+
+	if (diff->box.border.width != w->box.border.width ||
+	    diff->box.border.height != w->box.border.height) {
+		e.target = w;
+		e.data = NULL;
+		e.type = LCUI_WEVENT_RESIZE;
+		e.cancel_bubble = TRUE;
+		Widget_TriggerEvent(w, &e, NULL);
+		Widget_PostSurfaceEvent(w, LCUI_WEVENT_RESIZE,
+					!w->task.skip_surface_props_sync);
+		w->task.skip_surface_props_sync = TRUE;
+		Widget_AddTask(w, LCUI_WTASK_POSITION);
+		diff->invalid_box = SV_GRAPH_BOX;
+	}
+	if (diff->position != style->position) {
+		if (w->parent && style->position != SV_ABSOLUTE) {
+			Widget_AddTask(w->parent, LCUI_WTASK_LAYOUT);
+		}
+	}
+	if (MEMCMP(&diff->box.padding, &w->box.padding)) {
+		diff->invalid_box = max(diff->invalid_box, SV_PADDING_BOX);
+		Widget_UpdateChildrenSize(w);
+		Widget_AddTask(w, LCUI_WTASK_LAYOUT);
+	}
+	if (!diff->should_add_invalid_area) {
+		return 0;
+	}
+
+	/* Check for canvas related property changes */
+
+	if (diff->invalid_box == SV_GRAPH_BOX) {
+	} else if (diff->visible != style->visible) {
+		diff->invalid_box = SV_GRAPH_BOX;
+		if (style->visible) {
+			Widget_PostSurfaceEvent(w, LCUI_WEVENT_SHOW, TRUE);
+		} else {
+			Widget_PostSurfaceEvent(w, LCUI_WEVENT_HIDE, TRUE);
+		}
+	} else if (diff->z_index != style->z_index &&
+		   style->position != SV_STATIC) {
+		diff->invalid_box = SV_GRAPH_BOX;
+	} else if (MEMCMP(&diff->shadow, &style->shadow)) {
+		diff->invalid_box = SV_GRAPH_BOX;
+	} else if (diff->invalid_box == SV_BORDER_BOX) {
+	} else if (MEMCMP(&diff->border, &style->border)) {
+		diff->invalid_box = SV_BORDER_BOX;
+	} else if (MEMCMP(&diff->background, &style->background)) {
+		diff->invalid_box = SV_BORDER_BOX;
+	} else {
+		return 0;
+	}
+
+	/* Processing invalid area */
+
+	if (widget_box_moved && w->parent) {
+		if (!LCUIRectF_IsCoverRect(&diff->box.canvas, &w->box.canvas)) {
+			Widget_InvalidateArea(w->parent, &diff->box.canvas,
+					      SV_PADDING_BOX);
+			Widget_InvalidateArea(w, NULL, diff->invalid_box);
+			return 1;
+		}
+		LCUIRectF_MergeRect(&rect, &diff->box.canvas, &w->box.canvas);
+		Widget_InvalidateArea(w->parent, &rect, SV_PADDING_BOX);
+		return 1;
+	}
+	Widget_InvalidateArea(w, NULL, diff->invalid_box);
+	return 1;
 }
 
 LCUI_WidgetTaskContext Widget_BeginUpdate(LCUI_Widget w,
@@ -347,7 +925,8 @@ static size_t Widget_UpdateChildren(LCUI_Widget w, LCUI_WidgetTaskContext ctx)
 		DEBUG_MSG("%s %s: is visible\n", w->type, w->id);
 		if (data->rules.first_update_visible_children) {
 			total += Widget_UpdateVisibleChildren(w, ctx);
-			DEBUG_MSG("first update visible children count: %zu\n",
+			DEBUG_MSG("first update visible children "
+				  "count: %zu\n",
 				  total);
 		}
 	}
@@ -418,50 +997,68 @@ static size_t Widget_UpdateChildren(LCUI_Widget w, LCUI_WidgetTaskContext ctx)
 	return total;
 }
 
-size_t Widget_UpdateWithContext(LCUI_Widget w, LCUI_WidgetTaskContext ctx)
+static void Widget_UpdateSelf(LCUI_Widget w, LCUI_WidgetTaskContext ctx)
 {
 	int i;
-	size_t count = 0;
+	int self_updates;
 	LCUI_BOOL *states;
-	LCUI_WidgetTaskContext self_ctx;
 
-	if (!w->task.for_self && !w->task.for_children) {
-		return 0;
+	Widget_BeginDiff(w, ctx);
+	states = w->task.states;
+	/* 如果有用户自定义任务 */
+	if (states[LCUI_WTASK_USER] && w->proto && w->proto->runtask) {
+		states[LCUI_WTASK_USER] = FALSE;
+		w->proto->runtask(w);
+		if (ctx->profile) {
+			ctx->profile->user_task_count += 1;
+		}
 	}
-	self_ctx = Widget_BeginUpdate(w, ctx);
-	if (w->task.for_self) {
+	if (ctx->profile) {
+		if (states[LCUI_WTASK_REFRESH_STYLE]) {
+			ctx->profile->refresh_count += 1;
+		}
+		if (states[LCUI_WTASK_LAYOUT]) {
+			ctx->profile->layout_count += 1;
+		}
+		ctx->profile->update_count += 1;
+	}
+	for (self_updates = 0; w->task.for_self && self_updates < 4;
+	     ++self_updates) {
 		w->task.for_self = FALSE;
-		states = w->task.states;
-		/* 如果有用户自定义任务 */
-		if (states[LCUI_WTASK_USER] && w->proto && w->proto->runtask) {
-			w->proto->runtask(w);
-			if (self_ctx->profile) {
-				self_ctx->profile->user_task_count += 1;
-			}
-		}
-		if (self_ctx->profile) {
-			if (states[LCUI_WTASK_REFRESH_STYLE]) {
-				self_ctx->profile->refresh_count += 1;
-			}
-			if (states[LCUI_WTASK_LAYOUT]) {
-				self_ctx->profile->layout_count += 1;
-			}
-			self_ctx->profile->update_count += 1;
-		}
 		for (i = 0; i < LCUI_WTASK_USER; ++i) {
 			if (states[i]) {
 				states[i] = FALSE;
 				if (self.handlers[i]) {
 					self.handlers[i](w);
 				}
-			} else {
-				states[i] = FALSE;
 			}
 		}
-		Widget_AddState(w, LCUI_WSTATE_UPDATED);
-		count += 1;
 	}
-	count += Widget_UpdateChildren(w, self_ctx);
+	if (self_updates >= 4) {
+		Logger_Warning("%s widget has too many duplicate updates\n",
+			       w->type);
+	}
+	Widget_EndDiff(w, ctx);
+	Widget_AddState(w, LCUI_WSTATE_UPDATED);
+}
+
+static size_t Widget_UpdateWithContext(LCUI_Widget w,
+				       LCUI_WidgetTaskContext ctx)
+{
+	size_t count = 0;
+	LCUI_WidgetTaskContext self_ctx;
+
+	if (!w->task.for_self && !w->task.for_children) {
+		return 0;
+	}
+	self_ctx = Widget_BeginUpdate(w, ctx);
+	Widget_InitDiff(w, self_ctx);
+	if (w->task.for_self) {
+		Widget_UpdateSelf(w, self_ctx);
+	}
+	if (w->task.for_children) {
+		count += Widget_UpdateChildren(w, self_ctx);
+	}
 	Widget_SortChildrenShow(w);
 	Widget_EndUpdate(self_ctx);
 	return count;
@@ -474,20 +1071,26 @@ size_t Widget_Update(LCUI_Widget w)
 
 size_t LCUIWidget_Update(void)
 {
-	size_t i, count;
+	unsigned i;
+	size_t count;
 	LCUI_Widget root;
+	const LCUI_MetricsRec *metrics;
 
-	/* 前两次更新需要主动刷新所有部件的样式，主要是为了省去在应用程序里手动调用
-	 * LCUIWidget_RefreshStyle() 的麻烦 */
-	if (self.update_count < 2) {
+	metrics = LCUI_GetMetrics();
+	if (memcmp(metrics, &self.metrics, sizeof(LCUI_MetricsRec))) {
+		self.refresh_all = TRUE;
+	}
+	if (self.refresh_all) {
 		LCUIWidget_RefreshStyle();
-		self.update_count += 1;
 	}
 	root = LCUIWidget_GetRoot();
-	for (count = i = 0; i < self.max_updates_per_frame; ++i) {
-		count = Widget_Update(root);
+	for (count = i = 0; i < 4; ++i) {
+		count += Widget_Update(root);
 	}
+	root->state = LCUI_WSTATE_NORMAL;
 	LCUIWidget_ClearTrash();
+	self.metrics = *metrics;
+	self.refresh_all = FALSE;
 	return count;
 }
 
@@ -503,20 +1106,26 @@ void Widget_UpdateWithProfile(LCUI_Widget w, LCUI_WidgetTasksProfile profile)
 
 void LCUIWidget_UpdateWithProfile(LCUI_WidgetTasksProfile profile)
 {
-	size_t i;
+	unsigned i;
 	LCUI_Widget root;
+	const LCUI_MetricsRec *metrics;
 
 	profile->time = clock();
-	/* 初次更新时主动刷新所有部件的样式，为了省去在应用程序里手动调用
-	 * LCUIWidget_RefreshStyle() 的麻烦 */
-	if (self.update_count < 1) {
+	metrics = LCUI_GetMetrics();
+	if (memcmp(metrics, &self.metrics, sizeof(LCUI_MetricsRec))) {
+		self.refresh_all = TRUE;
+	}
+	if (self.refresh_all) {
 		LCUIWidget_RefreshStyle();
-		self.update_count += 1;
+	}
+	if (self.refresh_all) {
+		LCUIWidget_RefreshStyle();
 	}
 	root = LCUIWidget_GetRoot();
-	for (i = 0; i < self.max_updates_per_frame; ++i) {
+	for (i = 0; i < 4; ++i) {
 		Widget_UpdateWithProfile(root, profile);
 	}
+	root->state = LCUI_WSTATE_NORMAL;
 	profile->time = clock() - profile->time;
 	profile->destroy_time = clock();
 	profile->destroy_count = LCUIWidget_ClearTrash();
