+++ test/test_widget_rect.c
@@ -1,4 +1,6 @@
-﻿#include <string.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
 #include <LCUI_Build.h>
 #include <LCUI/LCUI.h>
 #include <LCUI/timer.h>
